chore: refactor worker pool to use single pool
