+++ platform/src/apis/Platform.Api.ChartRendering/src/functions/shared/pool.ts
@@ -1,0 +1,7 @@
+import Piscina from "piscina";
+
+const piscina = new Piscina({
+  filename: "./dist/src/functions/shared/worker.js",
+});
+
+export default piscina;
